split record components into seperate smaller components
