+++ src/components/record/RecommendRecordList.tsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const RecommendRecordList = () => {
+    return (
+        <div>
+            
+        </div>
+    )
+}
+
+export default RecommendRecordList;
