+++ src/components/record/RecordListByCategory.tsx
@@ -1,20 +1,17 @@
 import React, {useEffect} from 'react';
-import { List, Card } from 'antd';
 import axios from 'axios';
-import Meta from 'antd/lib/card/Meta';
-import { ShoppingCartOutlined, HeartOutlined } from '@ant-design/icons';
-import { RecordModel } from "./interface/RecordModel";
-import { StatusCodeModel } from "./interface/StatusCodeModel";
-import { Link } from "react-router-dom";
-import Loading from "./warning/Loading";
-import Error from "./warning/Error";
+import { RecordModel } from "../interface/RecordModel";
+import { StatusCodeModel } from "../interface/StatusCodeModel";
+import Loading from "../warning/Loading";
+import Error from "../warning/Error";
+import RecordList from "./RecordList";
 
 type Props = {
     name: string;
 }
 
-const RecordList = ({name}: Props) => {
-    const [records, setRecords] = React.useState<RecordModel[]>({} as RecordModel[]);
+const RecordListByCategory = ({name}: Props) => {
+    const [records, setRecords] = React.useState<RecordModel[]>([] as RecordModel[]);
     const [loading, setLoading] = React.useState<boolean>(true);
     const [error, setError] = React.useState<boolean>(false);
     const [statusCode, setStatusCode] = React.useState<StatusCodeModel>({code: "500"});
@@ -61,28 +58,8 @@ const RecordList = ({name}: Props) => {
     }
 
     return (
-        loading ? <Loading size={35} /> : <List
-            grid={{
-                gutter: 16,
-                xs: 1,
-                sm: 2,
-                md: 4,
-                lg: 4,
-                xl: 6
-            }}
-            dataSource={records}
-            renderItem={item => (
-                <List.Item>
-                    <Card 
-                        cover={<img src={item.albumCover} alt="cover" />} 
-                        actions={[<ShoppingCartOutlined />, <HeartOutlined />]}
-                        loading={loading} >
-                            <Meta title={<Link to={"/products/" + item.name.split(" ").join("-").toLowerCase()}>{item.name}</Link>} description={"by " + item.artist + " for " + item.price + "$"} />
-                    </Card>
-                </List.Item>
-            )}
-        />
+        loading ? <Loading size={35} /> : <RecordList records={records} />
     )
 }
 
-export default RecordList;
+export default RecordListByCategory;
